Escape like characters from value
bindValue doesn't escape % or _. So if the method is LIKE or NOTLIKE, and
the value is a string, any of the % or _ chars are escaped.

It's important to note here that using a LIKE or NOTLIKE automatically
prepends and appends the value with %. This should be modified to give the
user more control over this parameter.