Moved insert from dynamo to pdoITable::insertDynamo
Dynamos keep track of their created pdoITable. That connection is used to
insert, update or load a dynamo using the dynamo's data. The method to
insert a dynamo is too large to be ok with it being in a delta method, so
I moved the method to the pdoITable and updated it to reflect that change.
Dynamo's insert method now just sends itself to the
pdoITable::insertDynamo method

Due to the way this fix was applied, other adjustments made thier way into
this commit. In general I cleaned up some variable names ($e, $t) =>
($dynamo, $pdoITable) as an example. In addition some commented out code
blocks and unused variables were removed.

toConsider: perhaps we should just take the database interaction methods
out of the delta methods and put them in a trait which the dynamo
implements? Dynamo's will still be able to take dynamic methods, and the
method to interact with a table should be seperate from the functionality
of the dynamo itself, the pdoITable can still be assigned to the dynamo at
creation.