Created phpDoc for sqlSpinner
sqlSpinner documentation has been updated to match the phpDoc spec.

Added some support for the UNION select capacity.

In the aggregate method, if no columns or alias are provided, the method
name is used as the alias.

Fixed some undefined variables. ($col was replaced with the array
position it was intending to reference, a for loop counter had a typo)

Tried to atomize the sql blocks. phpStorm didn't like the sql strings
ending with an empty space, so instead it's being concatenated between the
sql and the value. This seems to be visually cleaner anyway.

Some of the methods returns were inside of the try block instead of after.
This would create an issue with the spinner not always coming back.
If we want to die on error we should catch and throw errors.

The SET method now verifies that the spinner's state is 'update' before
appending the SET clause.

Removed some unused variables

Fixed some indentation