Prevent too many persistent connections
The cleanPDO test demonstrated that the issue causing the user test to
fail was all the times a new pdoITable (and thus cleanPDO with a
persistent connection) was created. By creating it once and managing the
relationships, we can prevent the issue from occuring.

OR we add 'driver_options'=>[persistent_connection=>false] to the config
object to disable the persistent connection for this test file.